extract useTimer function for turning end time into seconds
